requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-filter==25.1
django-stubs==5.2.1
django-stubs-ext==5.2.1
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
idna==3.10
Markdown==3.8.2
pillow==11.2.1
requests==2.32.4
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing_extensions==4.14.0
urllib3==2.5.0
